backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
supabase
python-jose[cryptography]
passlib[bcrypt]
openpyxl
python-dotenv